+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Add `optionalTableField`, `readOnlyTableField`,
+  `requiredTableField`, to replace `optional`, `readOnly` and
+  `required` in a later version.
+
 * Add `Opaleye.Unpackspec` as the forward-compatible place to import
   `Unpackspec` from